+++ examples/full_with_bloc/lib/screens/post_list.screen.dart
@@ -40,7 +40,7 @@ class _PostListScreenState extends State<PostListScreen> {
                         width: 20,
                         child: CircularProgressIndicator(
                           valueColor:
-                              AlwaysStoppedAnimation<Color>(Colors.white),
+                              AlwaysStoppedAnimation<Color>(Colors.black),
                         ),
                       ),
                     ),
@@ -68,7 +68,7 @@ class _PostListScreenState extends State<PostListScreen> {
             icon: const Icon(Icons.arrow_right_alt),
             onPressed: () =>
                 Navigator.pushReplacementNamed(context, JokeScreen.routeName),
-          )
+          ),
         ],
       ),
       body: BlocBuilder<PostBloc, PostState>(
@@ -110,7 +110,7 @@ class _PostListScreenState extends State<PostListScreen> {
                   padding: EdgeInsets.only(
                     bottom: MediaQuery.of(context).padding.bottom,
                   ),
-                )
+                ),
               ],
             );
           }